Bkk 11/06/67

+++ lib/views/widget/taskfrom/widget/task_edit_morevert.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:fristprofigmatest/colors/colors.dart';
 import 'package:fristprofigmatest/services/deletetask_serveice.dart';
+import 'package:fristprofigmatest/views/widget/taskfrom/task_list_page.dart';
 import 'package:get/get.dart';
 
-Future<bool?> showTaskEditBottomSheet(
-    BuildContext context, int taskId, String title, String desc) {
+Future<bool?> showTaskEditBottomSheet(BuildContext context, int taskId,
+    String title, String desc, bool completed) {
   return showModalBottomSheet<bool>(
     context: context,
     shape: const RoundedRectangleBorder(
@@ -33,6 +34,7 @@ Future<bool?> showTaskEditBottomSheet(
                     'taskId': taskId.toString(),
                     'title': title,
                     'desc': desc,
+                    'completed': completed.toString(),
                   },
                 );
                 Get.back(result: result);
@@ -49,9 +51,50 @@ Future<bool?> showTaskEditBottomSheet(
                 style: TextStyle(color: NotificationColors.dark, fontSize: 16),
               ),
               onTap: () async {
-                print(taskId);
-                await DeleteTaskService.deleteTask(taskId);
-                Get.back(result: true); // ปิด bottom sheet และส่งผลลัพธ์กลับ
+                bool confirmDelete = await Get.defaultDialog<bool>(
+                      title: "ยืนยันการลบ",
+                      middleText: "คุณต้องการลบ Todo นี้หรือไม่?",
+                      textConfirm: "ลบ",
+                      textCancel: "ยกเลิก",
+                      confirmTextColor: Colors.white,
+                      buttonColor: Colors.red,
+                      cancelTextColor: Colors.red,
+                      onConfirm: () {
+                        Get.back(result: true);
+                      },
+                      onCancel: () {
+                        Get.back(result: false);
+                      },
+                    ) ??
+                    false;
+
+                if (confirmDelete) {
+                  try {
+                    await DeleteTaskService.deleteTask(taskId);
+                    Get.snackbar(
+                      'ลบ Todo แล้ว',
+                      'ลบ Todo สำเร็จ',
+                      snackPosition: SnackPosition.TOP,
+                      backgroundColor: Colors.red,
+                      colorText: Colors.white,
+                    );
+                    Get.back(
+                        result: true); // ปิด bottom sheet และส่งผลลัพธ์กลับ
+                    Get.offAll(() => TaskListPage()); // กลับไปที่หน้า TaskList
+                  } catch (e) {
+                    Get.snackbar(
+                      'ลบ Todo ไม่สำเร็จ',
+                      'เกิดข้อผิดพลาดในการลบ Todo',
+                      snackPosition: SnackPosition.TOP,
+                      backgroundColor: Colors.red,
+                      colorText: Colors.white,
+                    );
+                    Get.back(
+                        result: false); // ปิด bottom sheet และส่งผลลัพธ์กลับ
+                  }
+                } else {
+                  Get.back(result: false); // ปิด bottom sheet หากยกเลิก
+                }
               },
             ),
           ],
